+++ belokurskaya.sofia/T3/commands.hpp
@@ -30,7 +30,7 @@ namespace belokurskaya
   double calculateAreaBasedOnSizeEven(const Polygon& polygon);
   double calculateAreaBasedOnSizeOdd(const Polygon& polygon);
   double calculateMeanArea(const std::vector< Polygon >& polygons, const Polygon& polygon);
-  double calculateAreaBasedOnVertexCount(const Polygon& polygon, int numVertexes);
+  double calculateAreaBasedOnVertexCount(const Polygon& polygon, size_t numVertexes);
   double sumPolygonAreas(const double sum, const Polygon& polygon, std::function< double(const Polygon&) > resultFunc);
   size_t compareNumVertexes(const Polygon& polygon, size_t numVertexes);
   double calcPolyCount(double sum, const Polygon& polygon, size_t(*resultFuncForCount)(const Polygon&));
